Create different useState for Search and Menu functionality
